% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinaR.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{annotatePeaks}
\usage{
annotatePeaks(cp, reference.genome, show.annotation.pie = FALSE, verbose)
}
\arguments{
\item{cp}{bed formatted consensus peak matrix: CHR, START, STOP and raw peak counts (peaks by 3+samples)}

\item{reference.genome}{genome of interested species. It should be 'hg38', 'hg19' or 'mm10'.}

\item{show.annotation.pie}{shows the annotation pie chart produced with ChipSeeker}

\item{verbose}{prints messages through running the pipeline}
}
\value{
DApeaks returns DA peaks
}
\description{
Runs DA pipeline and makes it ready for enrichment analyses
}
