% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{bed}
\alias{bed}
\title{Example peaks from bone marrow of B6 mice}
\format{
An object of class \code{data.frame} with 1000 rows and 25 columns.
}
\usage{
data(atac_seq_consensus_bm)
}
\description{
Example peaks from bone marrow of B6 mice
}
\examples{
data(atac_seq_consensus_bm)
}
\keyword{datasets}
