% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinaR.R
\name{cinaR}
\alias{cinaR}
\title{cinaR}
\usage{
cinaR(
  consensus.peaks,
  contrasts,
  DA.choice = 1,
  DA.fdr.threshold = 0.05,
  DA.lfc.threshold = 0,
  save.DA.peaks = F,
  DA.peaks.path = NULL,
  norm.method = "cpm",
  filter.method = "custom",
  library.threshold = 2,
  cpm.threshold = 1,
  TSS.threshold = 50000,
  show.annotation.pie = F,
  reference.genome = NULL
)
}
\arguments{
\item{consensus.peaks}{bed formatted consensus peak matrix: CHR, START, STOP and raw peak counts (peaks by 3+samples)}

\item{contrasts}{user-defined contrasts for comparing samples}

\item{DA.choice}{determines which pipeline to run:
(1) edgeR, (2) limma-voom, (3) limma-trend, (4) DEseq2.
Note: Use limma-trend if consensus peaks are already normalized, otherwise use other methods.}

\item{DA.fdr.threshold}{fdr cut-off for differential analyses}

\item{DA.lfc.threshold}{log-fold change cutoff for differential analyses}

\item{save.DA.peaks}{saves differentially accessible peaks to an excel file}

\item{DA.peaks.path}{the path which the excel file of the DA peaks will be saved,
if not set it will be saved to current directory.}

\item{norm.method}{normalization method for consensus peaks}

\item{filter.method}{filtering method for low expressed peaks}

\item{library.threshold}{number of libraries a peak occurs so that it is not filtered default set to 2}

\item{cpm.threshold}{count per million threshold for not to filter a peak}

\item{TSS.threshold}{Distance to transcription start site in base-pairs. Default set to 50,000.}

\item{show.annotation.pie}{shows the annotation pie chart produced with ChipSeeker}

\item{reference.genome}{genome of interested species. It should be 'hg38', 'hg19' or 'mm10'.}
}
\value{
DApeaks returns DA peaks
}
\description{
Runs DA pipeline and makes it ready for enrichment analyses
}
