% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_values.R
\docType{data}
\name{color_values}
\alias{color_values}
\title{color values}
\format{
An object of class \code{character} of length 8.
}
\usage{
color_values
}
\description{
color values
}
\keyword{datasets}
