% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinaR.R
\name{differentialAnalyses}
\alias{differentialAnalyses}
\title{Differential Analyses}
\usage{
differentialAnalyses(
  cp,
  contrasts,
  DA.choice,
  DA.fdr.threshold,
  DA.lfc.threshold,
  save.DA.peaks
)
}
\arguments{
\item{cp}{Consensus peaks}

\item{contrasts}{user-defined contrasts for comparing samples}

\item{DA.choice}{determines which pipeline to run:
(1) edgeR, (2) limma-voom, (3) limma-trend, (4) DEseq2}

\item{DA.fdr.threshold}{fdr cut-off for differential analyses}

\item{DA.lfc.threshold}{log-fold change cutoff for differential analyses}

\item{save.DA.peaks}{saves differentially accessible peaks to an excel file}
}
\value{
DApeaks returns DA peaks
}
\description{
Runs differential analyses pipeline of choice on consensus peaks
}
