% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinaR.R
\name{filterConsensus}
\alias{filterConsensus}
\title{filterConsensus}
\usage{
filterConsensus(
  cp,
  filter.method = "custom",
  library.threshold = 2,
  cpm.threshold = 1
)
}
\arguments{
\item{cp}{bed formatted consensus peak matrix: CHR, START, STOP and raw peak counts (peaks by 3+samples)}

\item{filter.method}{filtering method for low expressed peaks}

\item{library.threshold}{number of libraries a peak occurs so that it is not filtered default set to 2}

\item{cpm.threshold}{count per million threshold for not to filter a peak}
}
\value{
DApeaks returns DA peaks
}
\description{
Filters lowly expressed peaks from down-stream analyses
}
