% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{grch37}
\alias{grch37}
\title{Grch37}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66978 rows and 3 columns.
}
\usage{
data(grch37)
}
\description{
Grch37
}
\keyword{datasets}
