% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{grch38}
\alias{grch38}
\title{Grch38}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67495 rows and 3 columns.
}
\usage{
data(grch38)
}
\description{
Grch38
}
\keyword{datasets}
