% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{grcm38}
\alias{grcm38}
\title{Grcm38}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56597 rows and 3 columns.
}
\usage{
data(grcm38)
}
\description{
Grcm38
}
\keyword{datasets}
