% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{heatmap_plot}
\usage{
heatmap_plot(results)
}
\arguments{
\item{results}{cinaR result object}
}
\value{
ggplot object, pca plot
}
\description{
heatmap_plot
}
