% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{heatmap_var_peaks}
\alias{heatmap_var_peaks}
\title{heatmap_var_peaks}
\usage{
heatmap_var_peaks(results, heatmap.peak.count = 100, ...)
}
\arguments{
\item{results}{cinaR result object}

\item{heatmap.peak.count}{number of peaks to be plotted.
If number of peaks are less than k then all peaks will be used.}

\item{...}{additional arguments for heatmap function, for more info `?pheatmap`}
}
\value{
ggplot object
}
\description{
plot most variable k peaks (default k = 100) among all samples
}
\examples{
library(cinaR)
data(atac_seq_consensus_bm) # calls 'bed'

# creating dummy results
results <- NULL
results[["cp"]] <- bed[,c(4:25)]

heatmap_var_peaks(results)

}
