% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mouse2human}
\alias{mouse2human}
\title{mouse2human}
\usage{
mouse2human(genes)
}
\arguments{
\item{genes}{mice genes to be converted to human}
}
\value{
returns a mapping from mouse to human
}
\description{
mouse2human
}
