% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinaR.R
\name{normalizeConsensus}
\alias{normalizeConsensus}
\title{normalizeConsensus}
\usage{
normalizeConsensus(cp, norm.method = "cpm", log.option = FALSE)
}
\arguments{
\item{cp}{bed formatted consensus peak matrix: CHR, START, STOP and raw peak counts (peaks by 3+samples)}

\item{norm.method}{normalization method for consensus peaks}

\item{log.option}{logical, log option for cpm function in edgeR}
}
\value{
Normalized consensus peaks
}
\description{
Normalizes consensus peak using different methods
}
