% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{pca_plot}
\alias{pca_plot}
\title{pca_plot}
\usage{
pca_plot(results, overlaid.info, sample.names = NULL, show.names = TRUE)
}
\arguments{
\item{results}{cinaR result object}

\item{overlaid.info}{overlaid information onto the samples}

\item{sample.names}{names of the samples shown on pca plot}

\item{show.names}{logical, if set FALSE sample names will be hidden}
}
\value{
ggplot object, pca plot
}
\description{
pca_plot
}
