% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_rows}
\alias{scale_rows}
\title{scale_rows}
\usage{
scale_rows(x)
}
\arguments{
\item{x}{a matrix, possibly containing gene by samples}
}
\value{
Row-normalized matrix
}
\description{
Normalize (z-score) rows of a matrix
}
\examples{
\donttest{
library(cinaR)
data(atac_seq_consensus_bm) # calls 'bed'
bed.row.normalized <- scale_rows(bed[,c(4:25)])
head(bed.row.normalized)
}
}
