% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_comparisons}
\alias{show_comparisons}
\title{show_comparisons}
\usage{
show_comparisons(results)
}
\arguments{
\item{results}{output of the cinaR}
}
\value{
comparisons created
}
\description{
returns the names of the created comparisons
}
