% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verboseFn}
\alias{verboseFn}
\title{verboseFn}
\usage{
verboseFn(verbose)
}
\arguments{
\item{verbose}{boolean, determines whether the output going be printed or not}
}
\value{
print function
}
\description{
returns a printing function to be used with in the script
}
