% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{vp2008}
\alias{vp2008}
\title{Immune modules}
\format{
An object of class GMT; see \code{read.gmt} from qusage package.
}
\usage{
data(VP2008)
}
\description{
Immune modules
}
\references{
Chaussabel et al. (2008) Immunity 29:150-164
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2727981/}{PubMed})
}
\keyword{datasets}
