% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mouse2human}
\alias{mouse2human}
\title{mouse2human}
\usage{
mouse2human(genes)
}
\arguments{
\item{genes}{mice genes to be converted to human}
}
\value{
returns a mapping from mouse to human
}
\description{
Given the mice gene symbols, this function creates a map from mice to human
using biomaRt.
}
\examples{
\donttest{

mouse.genes <- c("Gfap", "Gzmb", "Il1b")
map <- mouse2human(mouse.genes)

}
}
