% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{run_enrichment}
\alias{run_enrichment}
\title{run_enrichment}
\usage{
run_enrichment(
  results,
  geneset = NULL,
  experiment.type = "ATAC-Seq",
  reference.genome = NULL,
  enrichment.method = NULL,
  enrichment.FDR.cutoff = 1,
  background.genes.size = 20000
)
}
\arguments{
\item{results}{list, DA peaks list for different contrasts}

\item{geneset}{Pathways to be used in enrichment analyses. If not set vp2008 (Chaussabel, 2008)
immune modules will be used. This can be set to any geneset using `read.gmt` function from `qusage`
package. Different modules are available: https://www.gsea-msigdb.org/gsea/downloads.jsp.}

\item{experiment.type}{The type of experiment either set to "ATAC-Seq" or "RNA-Seq"}

\item{reference.genome}{genome of interested species. It should be 'hg38', 'hg19' or 'mm10'.}

\item{enrichment.method}{There are two methodologies for enrichment analyses,
Hyper-geometric p-value (HPEA) or Geneset Enrichment Analyses (GSEA).}

\item{enrichment.FDR.cutoff}{FDR cut-off for enriched terms, p-values
are corrected by Benjamini-Hochberg procedure}

\item{background.genes.size}{number of background genes for hyper-geometric p-value
calculations. Default is 20,000.}
}
\description{
run_enrichment
}
\examples{
## For ATAC-Seq experiments
## run_enrichment(results)

## For RNA-Seq experiments
## run_enrichment(results, experiment.type = "RNA-Seq")
}
